+++ test/debian/molecule/default/tests/test_testaid_ansiblefacts.py
@@ -3,5 +3,5 @@ import testaid
 testinfra_hosts = testaid.hosts()
 
 
-def test_testaid_ansiblefacts_present(testvars):
-    assert 'ansible_distribution_version' in testvars['ansible_facts'].keys()
+def test_testaid_ansiblefacts_present(host, testvars):
+    assert 'ansible_distribution_version' in testvars['ansible_facts']