+++ test/debian/roles/anarchism/molecule/default/tests/test_anarchism_inventory.py
@@ -1,0 +1,7 @@
+import testaid
+
+testinfra_hosts = testaid.hosts()
+
+
+def test_anarchism_inventory(host, testvars):
+    assert testvars['group_vars_all'] == 'my_group_vars_value'